requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
shap==0.41.0
torch==1.12.1
tqdm==4.64.1
transformers==4.24.0
xgboost==1.7.6
